Several updates
Continued work on the view post and added more information when user clicks on a meme.
The user now will see posted by and posted date when clicking on a meme post.
Added a clickable link in the posted by so user can go to the post owners profile from the post the user opened.

 - Added some more styling to keep in line with the theme of the website.

Began working on the README for the project.